+++ sources/Booleans/ft_isalpha.c
@@ -1,48 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_is.c                                            :+:      :+:    :+:   */
+/*   ft_isalpha.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/08/29 14:04:10 by adantas-          #+#    #+#             */
-/*   Updated: 2023/01/28 12:53:11 by adantas-         ###   ########.fr       */
+/*   Created: 2023/04/25 17:23:15 by adantas-          #+#    #+#             */
+/*   Updated: 2023/04/25 17:23:51 by adantas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/libft.h"
+#include "../../includes/libft.h"
 
-int	ft_isalnum(int c)
-{
-	if (ft_isdigit(c) || ft_isalpha(c))
-		return (1);
-	return (0);
-}
-
+/**
+ * @brief Checks for an alphabetic character;
+ * 
+ * @param c The character to be checked;
+ * @return true if the character is an alphabetic character; otherwise, false;
+ */
 int	ft_isalpha(int c)
 {
 	if ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z'))
-		return (1);
-	return (0);
-}
-
-int	ft_isascii(int c)
-{
-	if (c >= 0 && c <= 127)
-		return (1);
-	return (0);
-}
-
-int	ft_isdigit(int c)
-{
-	if (c >= '0' && c <= '9')
-		return (1);
-	return (0);
-}
-
-int	ft_isprint(int c)
-{
-	if (c >= ' ' && c < 127)
-		return (1);
-	return (0);
+		return (true);
+	return (false);
 }
