+++ sources/Booleans/ft_isdigit.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_isdigit.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/04/25 17:21:17 by adantas-          #+#    #+#             */
+/*   Updated: 2023/04/25 17:22:56 by adantas-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/libft.h"
+
+/**
+ * @brief Checks for a digit;
+ * 
+ * @param c The character to be checked;
+ * @return true if the character is a digit from 0 to 9; otherwise, false;
+ */
+int	ft_isdigit(int c)
+{
+	if (c >= '0' && c <= '9')
+		return (true);
+	return (false);
+}
