+++ sources/Booleans/ft_islower.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_islower.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/04/25 17:35:06 by adantas-          #+#    #+#             */
+/*   Updated: 2023/04/25 17:35:36 by adantas-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/libft.h"
+
+/**
+ * @brief Checks for a lowercase character;
+ * 
+ * @param c The character to be checked;
+ * @return true if the character is lowercase; otherwise, false;
+ */
+int	ft_islower(int c)
+{
+	if (c >= 'a' && c <= 'z')
+		return (true);
+	return (false);
+}
