+++ sources/Booleans/ft_isupper.c
@@ -1,36 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_is_pt2.c                                        :+:      :+:    :+:   */
+/*   ft_isupper.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/01/28 12:53:25 by adantas-          #+#    #+#             */
-/*   Updated: 2023/01/30 11:16:45 by adantas-         ###   ########.fr       */
+/*   Created: 2023/04/25 17:36:45 by adantas-          #+#    #+#             */
+/*   Updated: 2023/04/25 17:37:13 by adantas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/libft.h"
+#include "../../includes/libft.h"
 
-int	ft_islower(int c)
-{
-	if (c >= 'a' && c <= 'z')
-		return (1);
-	return (0);
-}
-
-int	ft_isprime(size_t nb)
-{
-	if (nb == 2 || nb == 3 || nb == 5 || nb == 7)
-		return (1);
-	else if (nb % 2 == 0 || nb % 3 == 0 || nb % 5 == 0 || nb % 7 == 0)
-		return (0);
-	return (1);
-}
-
+/**
+ * @brief Checks for an uppercase letter;
+ * 
+ * @param c The character to be checked;
+ * @return true if the character is uppercase; otherwise, false;
+ */
 int	ft_isupper(int c)
 {
 	if (c >= 'A' && c <= 'Z')
-		return (1);
-	return (0);
+		return (true);
+	return (false);
 }
